fix: normalize last_version_run to Long to avoid ClassCastException on upgrade
Legacy builds stored last_version_run as Long via putLong in SharedPreferences.
SharedPreferencesMigration preserves the original primitive type, so for
installs that upgraded through the SharedPreferences -> DataStore refactor
the value lives in DataStore as Long. Reading it via intPreferencesKey
threw ClassCastException on startup in checkShouldShowChangeLog.

Switch LAST_VERSION_RUN to longPreferencesKey and add a one-shot
DataStore migration that rewrites any non-Long entry under "last_version_run"
as Long. Covers both the legacy Long case (no-op) and fresh installs from
recent rc builds that had stored it as Int.

+++ feature/settings/src/main/kotlin/com/kelsos/mbrc/feature/settings/data/SettingsManagerDataStore.kt
@@ -237,7 +237,7 @@ class SettingsManagerDataStore(
       preferences[PreferenceKeys.LAST_VERSION_RUN] ?: DefaultValues.LAST_VERSION_RUN
     }.first()
 
-    val currentVersion = appInfo.versionCode
+    val currentVersion = appInfo.versionCode.toLong()
 
     return if (lastVersionCode < currentVersion) {
       dataStore.edit { preferences ->
